move-kbm-under-hosts-resource
