+++ testdata/entities/hosts/list_kbm_drive_slots_all.json
