+++ testdata/entities/hosts/list_kbm_networks.json
