+++ testdata/entities/hosts/list_kbm_networks_all.json
