+++ testdata/entities/hosts/list_kbm_networks_pageview.txt
