+++ testdata/entities/hosts/list_kbm_networks_template.txt
