+++ testdata/entities/hosts/list_kbm_power_feeds.json
