+++ testdata/entities/hosts/set_kbm_power.json
