--- testdata/entities/kbm/list_nodes.json
@@ -1,38 +1,0 @@
-[
-    {
-        "id": "testId",
-        "kubernetes_cluster_id": "testId",
-        "kubernetes_cluster_node_id": "testId",
-        "kubernetes_cluster_node_number": 1,
-        "rack_id": "testId",
-        "type": "kubernetes_baremetal_node",
-        "title": "example.aa",
-        "location_id": 0,
-        "location_code": "test",
-        "status": "active",
-        "operational_status": "",
-        "power_status": "",
-        "configuration": "",
-        "private_ipv4_address": null,
-        "public_ipv4_address": "1.2.3.4",
-        "lease_start_at": "",
-        "scheduled_release_at": null,
-        "oob_ipv4_address": "",
-        "configuration_details": {
-            "ram_size": 2,
-            "server_model_id": 1,
-            "server_model_name": "server-model-123",
-            "public_uplink_id": 2,
-            "public_uplink_name": "Public 1 Gbps without redundancy",
-            "private_uplink_id": 3,
-            "private_uplink_name": "Private 1 Gbps without redundancy",
-            "bandwidth_id": 4,
-            "bandwidth_name": "20000 GB",
-            "operating_system_id": 5,
-            "operating_system_full_name": "CentOS 7 x86_64"
-        },
-        "labels": null,
-        "created_at": "2025-01-01T12:00:00Z",
-        "updated_at": "2025-01-01T12:00:00Z"
-    }
-]
